fixed the appearance of popular albums and related artists and replaced the map inside with existing components
