+++ src/routes/search/searchComponents/ArtistsResults.js
@@ -2,12 +2,14 @@ import React, { useContext } from 'react';
 import { NavLink, Outlet } from 'react-router-dom';
 import DisplayContext from '../../../context/DisplayContext.js';
 import style from '../../MusicBox.module.css';
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faHeart } from "@fortawesome/free-regular-svg-icons";
 
 const ArtistsResults = ({ artists }) => {
   const [state, dispatch] = useContext(DisplayContext);
 
   return (
-    <div className={style.main}>
+    <div className={style.main} style={{marginTop: 0, marginBottom: 0, paddingBottom: 0}}>
       <div>
         {artists && (
           <div>
@@ -29,7 +31,13 @@ const ArtistsResults = ({ artists }) => {
                     <div className={style.albumName} title={`${artist.name}`}>
                       {artist.name}
                     </div>
-                    <div className={style.artistName}>{artist.owner ? artist.owner.display_name : artist.type.toUpperCase()}</div>
+                    <div className={style.artistName}>
+                      {artist.followers.total.toLocaleString()}{" "}
+                      <FontAwesomeIcon
+                        icon={faHeart}
+                        style={{ color: "pink", marginLeft: "5px" }}
+                      />
+                    </div>
                   </NavLink>
                 );
               })}
